requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
mistralai==0.0.11
pydantic==2.5.0
aiohttp==3.9.1
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
email-validator==2.1.0
